Support multiple instances of same header.
Currently, passing the same header multiple times to a UserAgent method does not work as the header method of HTTP::Request returns a list of the value, which breaks the map operation builing the headers argument to AnyEvent::HTTP::http_request. This is fixed now by calling header in scalar context.

It’s not perfect as it will concatenate all the values of the header in a single one. However this is reasonable as this is supposed to have the same meaning for a web server and it seems that AnyEvent::HTTP does not support multiple instances of the same header currently in any case.