Add options in the constructors that can be passed to the http_request calls.
Currently options can only be passed to the AnyEvent::HTTP::http_request calls when using directly the AnyEvent::UserAgent::request method, and not the get/post/... methods. And there is no options to set these options in the constructor of the user agent for all methods. These points are fixed by this change.

The motivation for this change can be found in this ticket: https://rt.cpan.org/Ticket/Display.html?id=156970

It is unclear if the actual bug is in AnyEvent::HTTP or in AnyEvent::UserAgent, but being able to deactivate persistent connections is a simple workaround.